+++ liveavatar/models/wan/wan_2_2/modules/s2v/model_s2v.py
@@ -68,7 +68,7 @@ def rope_apply(x, grid_sizes, freqs, start=None):
     output = []
     for i, _ in enumerate(x):
         s = x.size(1)
-        x_i = torch.view_as_complex(x[i, :s].to(torch.float64).reshape(
+        x_i = torch.view_as_complex(x[i, :s].to(torch.float32).reshape(
             s, n, -1, 2))
         freqs_i = freqs[i, :s]
         # apply rotary embedding
@@ -86,7 +86,7 @@ def rope_apply_cond(x, grid_sizes, freqs, start=None):
     output = []
     for i, _ in enumerate(x):
         s = x.size(1)
-        x_i = torch.view_as_complex(x[i, :s].to(torch.float64).reshape(
+        x_i = torch.view_as_complex(x[i, :s].to(torch.float32).reshape(
             s, n, -1, 2))
         freqs_i = freqs[i, :s]
         # apply rotary embedding
